requirements.txt
darts==0.25.0
torch==2.2.0
backtrader==1.9.78.123
streamlit==1.28.0
langchain==0.1.5
openai==1.6.1
python-dotenv==1.0.0
pandas>=2.0.3
matplotlib==3.7.2
alpha_vantage==2.3.1
numpy>=1.26.0  # Updated for Python 3.12 compatibility
scikit-learn>=1.3.0
xgboost>=2.0.0  # Added explicit XGBoost dependency
ta==0.10.2
plotly==5.18.0
pytest==7.4.0
black==23.7.0
flake8==6.1.0